+++ plugins/plugin-chart-composed/src/plugin/configs/axis.ts
@@ -19,6 +19,7 @@
 import { t } from '@superset-ui/core';
 import { ControlPanelsContainerProps, formatSelectOptions } from '@superset-ui/chart-controls';
 import { Layout } from '../../components/utils';
+import { QueryMode } from '../utils';
 
 export const xAxisLabel = {
   name: 'x_axis_label',
@@ -52,7 +53,11 @@ export const y2AxisLabel = {
     description: t('Show second Y Axis Label in the chart'),
     visibility: ({ form_data }: ControlPanelsContainerProps) =>
       // @ts-ignore (update in package)
-      form_data.use_y2_axis && form_data?.metrics?.length > 1 && form_data.layout === Layout.horizontal,
+      form_data.use_y2_axis &&
+      // @ts-ignore
+      form_data?.metrics?.length > 1 &&
+      form_data.layout === Layout.horizontal &&
+      form_data.query_mode !== QueryMode.raw,
   },
 };
 
@@ -81,6 +86,22 @@ export const yAxisTickLabelAngle = {
     choices: formatSelectOptions(['0', '45', '90']),
     default: '0',
     description: t('Set Y axis tick label angle in the chart'),
+  },
+};
+
+export const xAxisInterval = {
+  name: 'x_axis_interval',
+  config: {
+    freeForm: true,
+    type: 'SelectControl',
+    clearable: true,
+    label: t('X Axis Interval'),
+    renderTrigger: true,
+    choices: formatSelectOptions(['preserveStart', 'preserveEnd', 'preserveStartEnd']),
+    default: '',
+    description: t(
+      'If set 0, all the ticks will be shown. If set preserveStart", "preserveEnd" or "preserveStartEnd", the ticks which is to be shown or hidden will be calculated automatically.',
+    ),
   },
 };