+++ plugins/plugin-chart-composed/src/plugin/configs/orderBy.ts
@@ -18,7 +18,7 @@
  */
 import { QueryFormData, t } from '@superset-ui/core';
 import { ControlConfig, SelectControlConfig } from '@superset-ui/chart-controls/lib/types';
-import { MAX_FORM_CONTROLS, SortingType, SortingTypeNames } from '../utils';
+import { MAX_FORM_CONTROLS, QueryMode, SortingType, SortingTypeNames } from '../utils';
 
 type Sorting = [
   { name: string; config: ControlConfig<'CheckboxControl'> },
@@ -27,6 +27,7 @@ type Sorting = [
 
 const orderByMetric: Sorting[] = [];
 const orderByGroupBy: Sorting[] = [];
+const orderByColumns: Sorting[] = [];
 
 const getOrderByRow = (source: string, name: string, title: string, index?: number): Sorting => [
   {
@@ -43,7 +44,9 @@ const getOrderByRow = (source: string, name: string, title: string, index?: numb
         }`,
       ),
       visibility: ({ form_data }: { form_data: QueryFormData }) =>
-        index === undefined || !!form_data?.[source]?.[index],
+        ((form_data.query_mode === QueryMode.aggregate && (source === 'metrics' || source === 'groupby')) ||
+          (form_data.query_mode === QueryMode.raw && (source === 'y_column' || source === 'x_column'))) &&
+        (index === undefined || !!form_data[source]?.[index]),
     },
   },
   {
@@ -57,17 +60,24 @@ const getOrderByRow = (source: string, name: string, title: string, index?: numb
         label: SortingTypeNames[key],
       })),
       visibility: ({ form_data }: { form_data: QueryFormData }) =>
-        !!(form_data[`use_order_by_${name}_${index ?? 0}`] && (index === undefined || form_data?.[source]?.[index])),
+        ((form_data.query_mode === QueryMode.aggregate && (source === 'metrics' || source === 'groupby')) ||
+          (form_data.query_mode === QueryMode.raw && (source === 'y_column' || source === 'x_column'))) &&
+        !!(form_data[`use_order_by_${name}_${index ?? 0}`] && (index === undefined || form_data[source]?.[index])),
       default: SortingType.ASC,
       description: t(`Set Ascending / Descending sorting for ${title} ${index === undefined ? '' : index + 1}`),
     },
   },
 ];
 
-orderByMetric.push(getOrderByRow('metrics', 'metric', t('"metric"')));
-
 for (let i = 0; i < MAX_FORM_CONTROLS; i++) {
   orderByGroupBy.push(getOrderByRow('groupby', 'group_by', t('"group by"'), i));
 }
 
-export { orderByMetric, orderByGroupBy };
+for (let i = 0; i < MAX_FORM_CONTROLS; i++) {
+  orderByMetric.push(getOrderByRow('metrics', 'metric', t('"metric"'), i));
+}
+
+orderByColumns.push(getOrderByRow('x_column', 'x_column', t('"X Column"')));
+orderByColumns.push(getOrderByRow('y_column', 'y_column', t('"Y Column"')));
+
+export { orderByMetric, orderByGroupBy, orderByColumns };